+++ fugue/src/main/java/com/atlassian/fugue/Functions.java
@@ -22,6 +22,7 @@ import java.lang.ref.ReferenceQueue;
 import java.lang.ref.WeakReference;
 import java.util.Arrays;
 import java.util.HashMap;
+import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.function.BiFunction;
@@ -29,6 +30,7 @@ import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.function.Supplier;
 
+import static com.atlassian.fugue.Option.option;
 /**
  * Utility methods for Functions
  * <P>
@@ -250,7 +252,7 @@ public class Functions {
     }
 
     @Override public Option<B> apply(A a) {
-      return (p.test(a)) ? Option.option(f.apply(a)) : Option.<B> none();
+      return (p.test(a)) ? option(f.apply(a)) : Option.<B> none();
     }
 
     @Override public String toString() {
@@ -416,7 +418,7 @@ public class Functions {
    * @since 2.0
    */
   public static <A, B> Function<A, Option<B>> mapNullToOption(Function<? super A, ? extends B> f) {
-    return Functions.compose(Functions.<B> nullToOption(), f);
+    return Functions.compose(Functions.<B>nullToOption(), f);
   }
 
   /**
@@ -520,6 +522,37 @@ public class Functions {
   public static <A, B> Function<A, B> constant(final B constant) {
     return from -> constant;
   }
+
+  /**
+   * Create a function that performs a map lookup returning None for null
+   *
+   * If you do not need a nondefaulted return result using a method reference is preferred
+   * {@literal map::get}
+   * @param map map to use for lookup
+   * @param <A> map key type
+   * @param <B> map value type
+   * @return result of calling Map#get replacing null with none
+   * @see Functions#forMapWithDefault to supply a default value for none
+   */
+  public static <A,B> Function<A, Option<B>> forMap(final Map<A, B> map){
+    return a -> option(map.get(a));
+  }
+
+  /**
+   * Create a function that performs a map lookup supplying a default value when
+   * a Map#get returns null
+   *
+   * If you do not need a defaulted return result using a method reference is preferred
+   * {@literal map::get}
+   * @param map map to use for lookup
+   * @param <A> map key type
+   * @param <B> map value type
+   * @return result of calling Map#get returning defaultValue instead if the result was null
+   */
+  public static <A,B> Function<A, B> forMapWithDefault(final Map<A, B> map, B defaultValue){
+    return forMap(map).andThen(o -> o.getOrElse(defaultValue));
+  }
+
 
   /**
    * Creates a stack of matcher functions and returns the first result that